environment/requirements.txt
ads1015==0.0.6
arandr==0.1.9
asn1crypto==0.24.0
astral==2.2
automationhat==0.2.0
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
configparser==3.5.0b2
cookies==2.2.1
cryptography==2.6.1
drumhat==0.1.0
entrypoints==0.3
enum34==1.1.6
envirophat==1.0.0
enviroplus==0.0.3
ExplorerHAT==0.4.2
Flask==1.0.2
font-roboto==0.0.1
fonts==0.0.3
fourletterphat==0.1.0
funcsigs==1.0.2
gpiozero==1.5.1
i2cdevice==0.0.7
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
ltr559==0.1.1
MarkupSafe==1.1.0
microdotphat==0.2.1
mock==2.0.0
monotonic==1.5
mote==0.0.4
motephat==0.0.3
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
paho-mqtt==1.5.1
pantilthat==0.0.7
pbr==4.2.0
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
pimoroni-bme280==0.0.2
ply==3.11
pms5003==0.0.5
prometheus-async==19.2.0
prometheus-client==0.8.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pygame==1.9.4.post1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pyOpenSSL==19.0.0
pyserial==3.4
pytz==2020.1
pyxdg==0.25
rainbowhat==0.1.0
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
smbus==1.1.post2
smbus2==0.3.0
sn3218==1.2.7
sounddevice==0.4.1
spidev==3.4
ST7735==0.0.4
touchphat==0.0.1
twython==3.7.0
unicornhathd==0.0.4
urllib3==1.24.1
waitress==1.4.4
Werkzeug==0.14.1
wrapt==1.12.1

